+++ NewPlatform.Flexberry.ORM.ODataService/Model/DataObjectEdmModel.cs
@@ -2,42 +2,53 @@
 {
     using Microsoft.Spatial;
     using System;
-    using Action = NewPlatform.Flexberry.ORM.ODataService.Functions.Action;
     using System.Collections;
     using System.Collections.Generic;
     using System.Diagnostics.Contracts;
     using System.Linq;
     using System.Reflection;
+    using System.Web.OData;
 
     using ICSSoft.STORMNET;
+
+    using NewPlatform.Flexberry.ORM.ODataService.Functions;
 
     using Microsoft.OData.Edm;
     using Microsoft.OData.Edm.Library;
     using Microsoft.OData.Edm.Library.Expressions;
     using Microsoft.OData.Edm.Library.Values;
 
-    using NewPlatform.Flexberry.ORM.ODataService.Functions;
-    using Microsoft.Practices.Unity;
-    using Microsoft.Practices.Unity.Configuration;
-    using System.Web.OData;
+    using ICSSoft.Services;
+
+    using Unity;
+
+    using Action = NewPlatform.Flexberry.ORM.ODataService.Functions.Action;
 
     /// <summary>
     /// EDM-модель, которая строится на основе сборок с объектами данных (унаследованными от <see cref="DataObject"/>).
     /// </summary>
     public class DataObjectEdmModel : EdmModel
     {
+        /// <summary>
+        /// Service to export data from ORM.
+        /// </summary>
         public IExportService ExportService { get; set; }
+
+        /// <summary>
+        /// Service to export data from ORM.
+        /// </summary>
+        public IODataExportService ODataExportService { get; set; }
 
         /// <summary>
         /// Ссылка на IDataObjectEdmModelBuilder.
         /// </summary>
         public IDataObjectEdmModelBuilder EdmModelBuilder { get; set; }
-        
+
         /// <summary>
         /// Имя свойства ключа.
         /// </summary>
         public string KeyPropertyName => _metadata.KeyPropertyName;
-        
+
         /// <summary>
         /// Описание свойства ключа.
         /// </summary>
@@ -81,12 +92,28 @@
         {
             Contract.Requires<ArgumentNullException>(metadata != null);
             EdmModelBuilder = edmModelBuilder;
-            var container = new UnityContainer();
+            var container = UnityFactory.GetContainer();
             if (container != null)
             {
-                container.LoadConfiguration();
                 if (container.IsRegistered<IExportService>("Export"))
+                {
                     ExportService = container.Resolve<IExportService>("Export");
+                }
+
+                if (container.IsRegistered<IExportService>())
+                {
+                    ExportService = container.Resolve<IExportService>();
+                }
+
+                if (container.IsRegistered<IODataExportService>("Export"))
+                {
+                    ODataExportService = container.Resolve<IODataExportService>("Export");
+                }
+
+                if (container.IsRegistered<IODataExportService>())
+                {
+                    ODataExportService = container.Resolve<IODataExportService>();
+                }
             }
 
             _metadata = metadata;
@@ -97,6 +124,7 @@
             RegisterMasters();
             RegisterDetails();
             RegisterGeoIntersectsFunction();
+            RegisterGeomIntersectsFunction();
         }
 
         private void BuildTypeHierarchy()
@@ -187,7 +215,7 @@
                 IEdmPrimitiveType edmPrimitiveType = EdmTypeMap.GetEdmPrimitiveType(propertyType);
                 if (edmPrimitiveType != null)
                 {
-                    EdmStructuralProperty edmProp = edmEntityType.AddStructuralProperty(GetEntityPropertName(propertyInfo) , edmPrimitiveType.PrimitiveKind);
+                    EdmStructuralProperty edmProp = edmEntityType.AddStructuralProperty(GetEntityPropertName(propertyInfo), edmPrimitiveType.PrimitiveKind);
                     this.SetAnnotationValue(edmProp, new ClrPropertyInfoAnnotation(propertyInfo));
                 }
             }
@@ -746,6 +774,17 @@
             var edmParameterType = EdmTypeMap.GetEdmPrimitiveType(typeof(Geography));
             edmFunction.AddParameter("geography1", new EdmPrimitiveTypeReference(edmParameterType, false));
             edmFunction.AddParameter("geography2", new EdmPrimitiveTypeReference(edmParameterType, false));
+        }
+
+        private void RegisterGeomIntersectsFunction()
+        {
+            EdmFunction edmFunction;
+            var edmReturnType = EdmTypeMap.GetEdmPrimitiveType(typeof(bool));
+            edmFunction = new EdmFunction("geom", "intersects", new EdmPrimitiveTypeReference(edmReturnType, false), false, null, true);
+            AddElement(edmFunction);
+            var edmParameterType = EdmTypeMap.GetEdmPrimitiveType(typeof(Geometry));
+            edmFunction.AddParameter("geometry1", new EdmPrimitiveTypeReference(edmParameterType, false));
+            edmFunction.AddParameter("geometry2", new EdmPrimitiveTypeReference(edmParameterType, false));
         }
     }
 }