+++ Tests/NewPlatform.Flexberry.ORM.ODataService.Tests/SelfHostBaseODataServiceIntegratedTest.cs
@@ -9,10 +9,10 @@
 
     using ICSSoft.STORMNET.Business;
 
-    using Microsoft.Practices.Unity;
-
     using NewPlatform.Flexberry.ORM.ODataService.Extensions;
-    
+
+    using Unity;
+    using Unity.AspNet.WebApi;
 
     /// <summary>
     /// Базовый класс для тестирования работы с данными через ODataService.